refactor: add useFetchRecentProjects hook
