+++ src/types/projects.ts
@@ -1,0 +1,7 @@
+export type ProjectProps = {
+  id: string | number;
+  name: string;
+  html_url: string;
+  description: string;
+  topics: string[];
+};
